gs-plugin-loader-sync: Stop using a separate GMainContext for setup
This fixes some leaks (spotted by asan) in the unit tests. These leaks
are a result of `GSource`s being added to the `GMainContext` during
setup, and remaining there after setup is complete (holding on to
memory), while the test then runs in the global default `GMainContext`.

By using the global default `GMainContext` to run setup synchronously,
those `GSource`s get added to the same `GMainContext` which is
subsequently used for the rest of the test, and hence they get an
opportunity to be dispatched or removed later on.

Arguably the other synchronous wrappers in `gs-plugin-loader-sync.c`
should use the global default `GMainContext` too, but that’s a change I
don’t have time to check and test right now. The only reason I’m making
the change for the setup function now is because the unit tests are
currently failing under asan.

Signed-off-by: Philip Withnall <pwithnall@gnome.org>