flatpak: Check for an uninitialised remote_name
`-Werror=maybe-uninitialized` suggested that `remote_name` could be used
uninitialised here. I don’t think it’s possible (there will always be an
operation set by flatpak which will cause it to be set), but let’s
initialise the variable to `NULL` and add an assertion to encode our
assumptions about it being set, to catch this potential error.

Signed-off-by: Philip Withnall <pwithnall@gnome.org>