build: Remove -Winline warning
It’s not actually that useful, as the GObject type declaration macros
declare inline functions which gcc then spams a lot of warnings about
choosing not to inline.

This wasn’t noticed previously because the amount of inlining depends on
optimisation levels; previously I was building with `-O0`, but now I’m
building with `-Og`. Apparently that changed things.

Signed-off-by: Philip Withnall <pwithnall@gnome.org>