meson: Bump to libadwaita 1.6.alpha
This is needed for `AdwButtonRow`, which will be used in an upcoming
commit.

libadwaita 1.6 will be available in GNOME 47, so it’s safe for us to
depend on it.

While it doesn’t require a GTK newer than 4.13.4, take the opportunity
here to bump the subproject version of GTK to the next stable version,
so that we’re not subprojecting an unstable version which may contain
bugs.