app-translation-dialog: Update parameters after the AdwDialog port
gs-app-translation-dialog is now an AdwDialog, so don't call it with
"gs_shell_modal_dialog_present" anymore.

Also, this dialog is still using default-width/height properties, but
AdwDialog doesn't support them. So remove these properties and use
content-width instead.

The default-height will be removed without a replacement to
content-height, as GtkInfoWindow won't have a fixed size in a following
commit. Not using content-height will allow the dialog to naturally
increase its height according to the contents length; this avoids
triggering a scrollbar that scrolls just a little bit if text length
is longer in other translations.