gs-appstream: Fails to read app 'scope' in some cases
The read of the silo filename and scope looked into a wrong place,
causing the apps to be set with an "unknown" scope, thus anything
matched them in the GsPlugin cache, regardless the look up could
be once for a 'system' and once for a 'user' scope. As an outcome,
the details page may not show all alternatives for the apps, when
for example the same named flatpak remote had been available both
as the system and as the user installation.