gs-plugin-icons: Propagate plugin's 'scale' property to the icon downloader
That to not shrink icons to 160 pixels when the scale factor is larger than one,
thus for example with the scale factor 2 the minimum icon size is 320, not 160px.
The GsPlugin:scale property can be modified during runtime, thus bind the plugin
property with the downloader property to use the up-to-date value for download.

Closes https://gitlab.gnome.org/GNOME/gnome-software/-/issues/2559