flatpak: Ensure gs_flatpak_file_to_app_ref() won't return cached app
The function is used to convert a file to an app, which sets the GsApp:local-file
property afterwards, but the cached apps cannot have the property set, because they
are persistent and nothing unsets the property. The following searches would take
such cached app and return it with the local-file property set, which might be
obsolete and unrelated to such search.