fedora-langpacks: Replace GS_PLUGIN_ACTION_GET_LANGPACKS with a list job in tests
This means the PackageKit plugin needs to be pulled into the
fedora-langpacks tests, though, to refine the search results. This makes
sense anyway, as it gives a more representative end-to-end test of how
the langpacks stuff is supposed to work on a production system: the
fedora-langpacks plugin lists the packages, and the PackageKit plugin
refines them.

Signed-off-by: Philip Withnall <pwithnall@gnome.org>

Helps: #1472