gs-app-translation-dialog: Fix dialog not shrinking properly
Both GtkBox and the GtkLabel inside the "Translation Website" button
have unnecessary margins that prevent the dialog from shrinking properly
when the parent window is small. Also, the button label doesn't wrap,
which also contribute to the shrinking problem.

To fix this, allow the button label to wrap and remove the margins.

Closes https://gitlab.gnome.org/GNOME/gnome-software/-/issues/2552