packagekit: Unset app as "installed", when its package is not installed
The appstream plugin marks an app as installed when it finds a .metainfo.xml
file with the same <id/> in the expected place, without knowing whether
the corresponding package is truly installed or not. That makes incorrect
impression about the installed state when two apps have the same <id/>,
but are different versions of the app (like when compiled with different
features, available in different repositories).

This change turns "installed" state into an "available" state when there are
not installed all packages for the app.

Closes https://gitlab.gnome.org/GNOME/gnome-software/-/issues/2541