Revert "general: Use AdwSpinner in the "Loading Updates" and "Loading App Details" pages"
This reverts commit c8dc241dbb8411154f8ffbaa52d35549b6723ba2.

It introduced an undocumented dependency on an unreleased version of
libadwaita (specifically
https://gitlab.gnome.org/GNOME/libadwaita/-/commit/19c71053). The need
to eliminate a little bit of boilerplate code on the status page is not
urgent enough to balance out the inconvenience of depending on a
bleeding-edge version of libadwaita and rebuilding the CI images to use
it.

These changes can land again once `AdwSpinner` is in a released version
of libadwaita.

See https://gitlab.gnome.org/GNOME/gnome-software/-/merge_requests/2083#note_2162213