gs-shell: Fix glitchy headerbar padding
GsUpdatesPausedBanner is currently a "top" type of child of
GsShell's AdwToolbarView. This is incorrect AdwBanner isn't
a supported widget for this case¹. Also, this creates a
little gap below the banner, so a CSS hack was introduced in
26e5728df5897dec00080c8642013b53988d5004 to workaround this
issue.

However, the workaround made the headerbar bottom padding to be
smaller. To fix all of this, move the banner to the content
widget instead.

¹ https://gnome.pages.gitlab.gnome.org/libadwaita/doc/main/class.ToolbarView.html#description
Closes https://gitlab.gnome.org/GNOME/gnome-software/-/issues/2634