gs-appstream: Treat <content_rating> with no <content_attribute> values as valid
<content_rating type="*"/> tag without specific <content_attribute>
values should be considered as all attributes with a value of
'AS_CONTENT_RATING_VALUE_NONE'.

Explanation:
------------

From https://github.com/hughsie/oars/blob/HEAD/specification/oars-1.1.md

If an attribute is not present in a content rating, its value must be
assumed to be none. If no content rating is provided for an
application at all, however, the value of all attributes (in any
version of OARS) must be assumed to be unknown.