gs-os-update-page: Verify downgrade is a downgrade
The version can sometimes end with a non-version string, common in both strings,
which can confuse the comparison, thus try without the suffix, if such exists
and is not a number.

For example: "2:2.1-61.fc40" and "2:2.1-61.1.fc40" is reported as a downgrade,
but without the common suffix ".fc40" it's correctly reported as an update.

Closes https://gitlab.gnome.org/GNOME/gnome-software/-/issues/2273