meson: include current dir as include directories for libgnome-software
Anything that uses the library as a dependency needs access to both of
those directories for headers, etc. Instead of adding the include to
every consumer, simply add it to the library as a dependency. In
addition to being less code, this fixes the use of gnome-software as a
subproject.