gs-loading-page: Stop progress bar pulsing when unmapped
It’s unlikely that the loading page would be the chosen app page, and be
showing its pulsing progress bar for a long time, if the user closes the
gnome-software app window. However, it is possible. If so, we don’t want
the process to be waking up every 50ms to pulse a progress bar which
isn’t visible, so temporarily stop pulsing while the page is unmapped.

Signed-off-by: Philip Withnall <pwithnall@gnome.org>

Helps: #2640