gs-upgrade-banner: Stop pulsing progress bar when unmapped
There’s no need to pulse the progress bar every 50ms when the UI isn’t
shown because the user has closed the window. This could happen while
they’re waiting for a long upgrade process to happen. While the rest of
the upgrade activity will probably prevent the machine going into a
particularly deep sleep state, we might as well not cause additional
context switches into the gnome-software process to update the state of
a progress bar which isn’t visible.

Signed-off-by: Philip Withnall <pwithnall@gnome.org>

Helps: #2640