packagekit: Set update version only when package is updateable
A PkPackage can have 'PK_INFO_ENUM_AVAILABLE' when there is an update
available to be installed or when the package is available remotely
but yet to be installed. So, set the update version of the app only in
the former case, and actual version in the later.

Refer https://gitlab.gnome.org/GNOME/gnome-software/-/issues/2575