shell: Port to AdwNavigationView
AdwLeaflet is deprecated and replaced with AdwNavigationView,
which also allows us to remove the GNU deprecation macros.
AdwNavigationView supports shortcuts and gestures automatically,
and also has a popover menu that displays the page hierarchy
when press and holding the back button.

This MR removes the ability to search in any subpage, as the
current code design was heavily influenced by the limitations
of AdwLeaflet, which interferes with the port.

Fixes https://gitlab.gnome.org/GNOME/gnome-software/-/issues/2052