gs-page: Fix "Uninstall [app]?" alert dialog not being adaptive
GtkWidget:width-request=360 breaks adaptiveness of the
"Uninstal [app]" alert dialog, preventing it to be
properly resized when the parent window is smaller than
360px. Also, GtkWidget:halign and GtkWidget:hexpand do
nothing, so they are unnecessary.

Also, remove AdwActionRow:title/subtitle-lines=1, as
this makes the row title and subtitle to ellipsize when
the dialog width is small, making the text unreadable.
Remove these properties to make the text wrap instead.

Closes https://gitlab.gnome.org/GNOME/gnome-software/-/issues/2713