debugging: Fix instruction for showing gnome-software journal logs
The `journalctl --identifier` argument filters based on the
`SYSLOG_IDENTIFIER` journal field, which gnome-software doesn’t actually
set. This instruction definitely worked when it was originally
documented, so I’m unsure what has changed (and haven’t investigated).

Instead, pass the path of the gnome-software binary to `journalctl`, and
allow it to use that to match on `_EXE` or `_COMM` journal fields, which
should always be set by systemd, and are trusted.

Signed-off-by: Philip Withnall <pwithnall@gnome.org>