gs-details-page: Use a weak ref to track whether review dialog is open
Rather than trying to match pointers — as Sid said in
https://gitlab.gnome.org/GNOME/gnome-software/-/merge_requests/1944#note_2258289,
pointer addresses can be reused.

This commit changes the code to no longer really care whether the dialog
is open, merely whether the object representing it still exists. If it
still exists, we can call methods on it to set the error text and hide
it, even if they actually don’t end up changing the UI.

Signed-off-by: Philip Withnall <pwithnall@gnome.org>