gs-remote-icon: Specify scale separately from maximum icon size
The scale is often known out-of-band (for example, from the `scale`
attribute of metainfo `<icon>` tags). This allows us to correctly set
the logical dimensions of a downloaded icon.

As a result of this change, more parts of the code need to become aware
of scale, preventing device and logical pixels being conflated all over
the place.

Signed-off-by: Philip Withnall <pwithnall@gnome.org>

Helps: #2559