gs-appstream: Prefer swcatalog over app-info
When the user upgrades the machine, or when for example PackageKit
is updated and switches the appstream data installation from
/var/cache/app-info/ to /var/cache/swcatalog/, then the app-info path
might not be cleared, causing stale data being saved there, while
the swcatalog path is properly updated.

Switch the order of the scanned directories, to have precedence the data
in the swcatalog over the app-info data, if both exist.