appstream: Catch on Appstream data creation race
This will help in situations when the Appstream data caches are empty
and the corresponding directories nonexistent. The race is that in case
for example the PackageKit is started due to the gnome-software request,
then before it downloads and installs the repositories' Appstream data,
the gnome-software's appstream plugin is already running and initialized,
ignoring any additions into these paths, thus even the PackageKit installs
the Appstream data, the gnome-software does not know about it either until
restarted or when there's a change in any of the existing directories it
already listens for the changes at.