gs-updates-section: Fix an assertion failure on cancelling a restart
`app` is always `NULL` since commit 47d79ede. Remove the cancellable
entirely, because we don’t want to be able to cancel cancellation of the
restart. The UI doesn’t support it, it doesn’t really make sense, and
it’s only likely to happen as a result of mistakenly reusing an
already-cancelled `GCancellable`, i.e. unintentionally.

Thanks to Sid for reporting and debugging this.

Signed-off-by: Philip Withnall <pwithnall@gnome.org>

Fixes: #2732