gs-app-reviews-dialog: Remove redundant branch
If the async operation fails, an error must always be set — this is
convention for how `GError` works
(https://docs.gtk.org/glib/error-reporting.html#rules-for-use-of-gerror),
and `GsOdrsProvider` does seem to follow that convention.

If it doesn’t (i.e. if this assertion fails), that’s a bug in
`GsOdrsProvider`.

Signed-off-by: Philip Withnall <pwithnall@gnome.org>