gs-app-details-page: Use GtkStack to show spinner separately from label
Rather than switching between the loading spinner and the update details
label by changing their visibility and/or text content to/from empty,
use a `GtkStack` to switch between the two. This allows us to use the
`[vh]homogeneous` properties to ensure both widgets are the same size,
and allows them both to be centred independently.

Signed-off-by: Philip Withnall <pwithnall@gnome.org>