gs-details-page: Query job manager lazily
If the app isn’t going to change, there’s no need to get the job
manager.

In particular, this helps when `self->app == app == NULL`, which happens
when shutting down early (`gnome-software --quit` when gnome-software
isn’t already running). In that situation, the job manager is not set up
yet, so getting it fails. Avoid that.

Signed-off-by: Philip Withnall <pwithnall@gnome.org>

Helps: #2579