ci: Add dist job to build a dist tarball
This is run on a schedule and all tags. Running it on a schedule means
we get to periodically test that it works, which is important because we
only get one chance at a release. Running it on a tag creates a release
and automatically uploads it to download.gnome.org using the release
service.

See https://gitlab.gnome.org/Teams/Websites/handbook.gnome.org/-/merge_requests/83/

Signed-off-by: Philip Withnall <pwithnall@gnome.org>