build: Bump Meson requirement to 1.0.1
This is needed for `meson dist` to work on the CI machines. I haven’t
debugged the old version doesn’t work — it somehow doesn’t detect the
git repository properly and errors out saying that only git and
Mercurial repositories are supported.

Meson 1.0.1 is still a relatively old version: it’s the one which is
currently shipped in Debian Stable. So it should be available pretty
much everywhere.

This includes a backport of commit
71ede7bed130dfcff388acbb73119e2c12ad8331 to fix the Debian CI image
build.

Signed-off-by: Philip Withnall <pwithnall@gnome.org>