fwupd: Updates are not downloaded/installed
The install_or_update_apps_impl() is called with `-1` for either `install_flags`
or `update_flags`, depending whether it's called for the update or for the install.
The `-1` transforms into `0xffffff`, thus it means all the flags are set, especially
those lower bits, which correspond to the "no-download" flag, which was checked
unconditionally regarding this `-1` value, thus the code behaved like there is
always one of them set to "no-download", thus it did not download anything.
Similarly when applying the update or using the "interactive" flag.

Closes https://gitlab.gnome.org/GNOME/gnome-software/-/issues/2724