ci: Ensure submodules are initialised in CI jobs
We copy the submodules into the git working directory manually in CI
jobs, to prevent them being downloaded over the network and ensure
they’re all the expected version so that CI jobs are comparable.

However, git stores metadata about the status of each submodule outside
of the submodule’s directory (in this case, in particular, outside of
`subprojects/gnome-pwa-list`). We need to explicitly initialise the git
submodules in order to update that metadata in the working directory.

This fixes the following warning during `meson dist`:
```
WARNING: Submodule 'subprojects/gnome-pwa-list' is not checked out and cannot be added to the dist
```
which was emitted by Meson even when the `subprojects/gnome-pwa-list`
submodule is present and populated. The gnome-pwa-list subproject is new
since we last did a major release, which is why this issue hasn’t come
up before.

Because the Meson warning is only a warning (and not an error) this
unfortunately went undetected during the 47.alpha and 47.alpha2
releases.

Inspired by a suggestion by Milan Crha
(https://gitlab.gnome.org/GNOME/gnome-software/-/merge_requests/2188#note_2315257).

Additionally, set the `GIT_SUBMODULE_STRATEGY` and `GIT_SUBMODULE_DEPTH`
variables to prevent GitLab trying to check out the submodules itself
(there’s no need for it to), and to limit its clone depth if it does.

Signed-off-by: Philip Withnall <pwithnall@gnome.org>
Fixes: #2748