build: Add gs-build-ident.h to libgnomesoftware_dep
It’s a generated header file, so if (for example) `src/gs-application.c`
(which includes it) is compiled before libgnomesoftware is fully built
(since it only needs to be fully built in time for linking), the header
will not be found.

Correctly list `gs-build-ident.h` in the sources for
`libgnomesoftware_dep` to fix this.

Should fix CI failure: https://gitlab.gnome.org/GNOME/gnome-software/-/jobs/4655185
```
../src/gs-application.c:24:10: fatal error: gs-build-ident.h: No such file or directory
   24 | #include "gs-build-ident.h"
      |          ^~~~~~~~~~~~~~~~~~
compilation terminated.
```

Signed-off-by: Philip Withnall <pwithnall@gnome.org>