fwupd: Do not use cache for historical updates
Historical updates are just about updates which happened, well, in the history.
These do not need to be cached, neither reused for new updates. It's safer
to not cache them, thus the app's UpdateID (a URL to the update file) is
not overwritten by later lookups (the worse case to have a repeated historical
update lookup after new updates being found for the app).