ci: Support publishing help translations
This means we need the `help` directory from the build directory, as we
need Meson to have run to have merged the translations into the Yelp
`.page` files.

Index generation is done via a Python script, which means we can use
already-existing code to get human readable names for the locales.

Signed-off-by: Philip Withnall <pwithnall@gnome.org>

Helps: #2498