gs-appstream: Do not merge into to-be-ignored component
This could happen when there had been two or more flatpak remotes
referencing the same source and one of them filtering the apps
with `noenumerate` flag or such, then the merge code could merge
the AppStream data into a node marked to be ignored and then mark
the source node to be ignored as well, which effectively hides
all the apps from that non-filtered remote. These remotes needed
to be part of the same installation.

Closes https://gitlab.gnome.org/GNOME/gnome-software/-/issues/2766