gs-search-page: Show spinner without delay if starting a new search
There’s been code in `GsSearchPage` for many years which delays showing
the spinner when starting a search, in order to avoid briefly showing
the spinner when updating a search.

For example, this means that if you search for ‘rail’, then it waits
250ms, then shows the spinner, then eventually shows the results. If you
then extend the search to ‘railway’, the existing results keep showing
for 250ms before the spinner is shown — if the new results are found
before that timeout is reached, the new results are shown directly.

This works great for when an existing search is extended. If starting a
new search from scratch, though, the search page will start showing the
‘Search for Apps’ page, then wait 250ms (because a new search will
likely take longer than this), then show the spinner, then show the
results once they’re found. This looks flickery.

There’s no point in showing ‘Search for Apps’ for 250ms when starting a
new search.

Signed-off-by: Philip Withnall <pwithnall@gnome.org>

Fixes: #2771