gs-search-page: Don’t briefly show old results when searching again
If a search is done, then the user switches to another page (e.g.
‘Explore’), then does another search via typeahead find, the results
from the first search will briefly be visible while the user is typing
their second search.

This is a bit confusing for the user, as they’ve already switched away
from those search results, and they are no longer accessible anywhere
(e.g. in the back stack).

So, add a new method to allow the shell to explicitly clear old search
results, and call that when changing mode in a way which clears UI
context.

Signed-off-by: Philip Withnall <pwithnall@gnome.org>