flatpak: Fix check for whether new permissions are required by an update
The `GS_APP_QUIRK_NEW_PERMISSIONS` quirk was only being set if the
update required more permissions *flags* than the existing version of
the app. This covers most use cases, but it ignores the case where the
static file permissions of the app have increased in scope.

Use the new `gs_app_permissions_is_empty()` method to encapsulate this
logic.

Signed-off-by: Philip Withnall <pwithnall@gnome.org>