tests: Add basic tests for GsAppPermissions
As a standalone object in libgnomesoftware, it’s an ideal candidate for
unit testing.

This commit mostly adds a test framework for the `lib/` directory.
Following commits will add more unit tests. This approach is taken for
the tests so we can start to move to a setup where there are more,
smaller, tests, which are easier to run separately and easier to find
things in. The current setup, where there’s one big ’self test’ per
directory, is quite hard to understand — is the ‘self test’ trying to be
an integration test, or unit tests? How do I run only some of the tests,
to check a particular bit of code?

The approach taken in this commit is inspired by what’s done in GLib,
which has extensive and successful test coverage.

The `template.test.in` and `template-tap.test.in` and related changes
are support for ‘installed tests’ (see
https://wiki.gnome.org/Initiatives/GnomeGoals/InstalledTests), which
allows the unit tests to be run post-installation on a system, against
the system libraries they’ll be used with in production. See that wiki
page for the benefits of this approach for one aspect of integration
testing.

Signed-off-by: Philip Withnall <pwithnall@gnome.org>