build: Split test envs so tests use correct directory paths
Use the approach from GLib, in that common test environment data is set
at the top level using Meson’s `add_test_setup()`, then we only need to
set the directory-specific environment in each `meson.build` file.

As a side effect, this fixes the `G_TEST_SRCDIR` and `G_TEST_BUILDDIR`
paths for each test to reference the correct subdirectories, and it adds
a couple of diagnostic environment variables for GLib, which might make
debugging a bit easier.

Signed-off-by: Philip Withnall <pwithnall@gnome.org>