gs-app-permissions: Fix contains_filesystem_*() checks
`bsearch()` passes pointers to the elements to compare to the comparison
function. When dealing with a pointer array, this means that the
comparison function receives a pointer-to-a-pointer. Unfortunately, the
existing comparison function was treating the inputs as single pointers,
so was comparing the needle path against the value of a pointer.
Thankfully most pointers have a nul byte in them somewhere, so this
avoided most unbounded reads.

Fix the search to use pointers-to-pointers. This means the comparison
function can be combined with the existing `qsort()` one.

Signed-off-by: Philip Withnall <pwithnall@gnome.org>