flatpak: Significantly improve static bus permissions handling
Previously, the `GsAppPermissions` code in the flatpak plugin
highlighted where apps were either:
 * requiring unfiltered access to either the session bus or system bus;
 * or requiring static permissions to talk to a few specific services
   which we know give a sandbox escape mechanism.

So this was a basic blocklist mechanism for bus policy.

This was insufficient, though — almost all services on the session bus,
and some on the system bus, were not designed to isolate untrusted
clients from each other or from user and system information. Before the
era of sandboxing, the security model on Linux was that any user process
could be assumed to have read access to most of the system, and
read/write access to the user’s home directory, and everything would be
vetted at compilation and packaging time. Services were designed within
this security model.

Unfortunately, moving to a sandboxed security model where user processes
are mutually distrustful means they can’t be trusted to access most
D-Bus services either.

Hence, we actually need an allowlist for bus policy, rather than a
blocklist.

So, change the bus policy code in the flatpak plugin to iterate through
each policy in the app’s keyfile, and work out whether it is:
 * A default policy, such as permission to talk to a portal or own the
   app’s well-known name on the session bus. If so, ignore it.
 * A known policy, such as a known sandbox escape. If so, add the
   relevant flag to the `GsAppPermissions`.
 * Other, such as permission to talk to a service which isn’t known to
   gnome-software. If so, add the bus policy to the `GsAppPermissions`
   so it can be flagged specifically to the user.

This means we guarantee to make a decision about every bus policy in the
app’s manifest, if the app has filtered access to a bus.

This is all described very well on man:flatpak-metadata(5), which this
implementation is based off. The `.Devel` handling is a
flathub/freedesktop convention, and was added on top of flatpak’s
defaults.

Signed-off-by: Philip Withnall <pwithnall@gnome.org>

Helps: #2760