gs-app-context-bar: Expose static bus policies in the UI
Use the new API in `GsAppPermissions` to expose details of the bus
policies required by an app to the user, highlighting which ones might
indicate the app is unsafe.

I’ve used the terminology ‘non-portal service’ here to try and explain
what a D-Bus service is in a way which relates its security model to
flatpak sandboxing.

Signed-off-by: Philip Withnall <pwithnall@gnome.org>

Fixes: #2760