ci: Rework pages CI job dependency to use needs keyword
In pipeline
https://gitlab.gnome.org/asiersarasua/gnome-software/-/pipelines/810511/,
for some reason, the `fedora-x86_64` job has not run (I cannot figure out
why). Even though the `pages` job depends on it. This means the `pages`
job is failing because it can’t get the build artifacts from the earlier
job.

Try switching from `dependencies` to `needs` to see if that forces the
`fedora-x86_64` job to run if `pages` does.

See https://dev.to/zavoloklom/gitlab-ci-needs-vs-dependencies-a-practical-guide-aoi

Signed-off-by: Philip Withnall <pwithnall@gnome.org>