packagekit: Skip sending proxy settings to daemon if they’re empty
This saves several D-Bus round-trips on startup, which were pointless if
the proxy settings were empty, as the default for PackageKit is to have
empty proxy settings.

This is particularly useful in unit tests on CI, as the CI runners don’t
have a logind session running, so calling `SetProxy()` on the PackageKit
daemon fails and emits a critical warning, which then aborts the test.
While we wait for this to be improved in PackageKit
(https://github.com/PackageKit/PackageKit/pull/850), this commit
provides a workaround for that too.

Signed-off-by: Philip Withnall <pwithnall@gnome.org>