rpm-ostree: Do not hard-code refspec for distro upgrade
The distro can be other than the Silverblue variant, thus rather derive
the new refspec from the booted deployment origin and change only the version
in it. That will avoid reset of the current variant for example from Kinoite
to Silverblue. It also ensures the expected refspec format is used.

Closes https://gitlab.gnome.org/GNOME/gnome-software/-/issues/2813