gs-appstream: Add thread safety into refine of relations
Even the relations array is reffed in the gs_app_get_relations(),
the gs_app_add_relation() modified the existing array during the refine,
thus when one thread uses the array, it could be modified from
another thread, which can lead to trouble.

This makes sure the array is not modified once it is returned from
the gs_app_get_relations() and adds thread safety by setting all
the relations within a single call.

Related to a downstream crash report:
https://bugzilla.redhat.com/show_bug.cgi?id=2364081