gs-plugin-job-file-to-app: Add a :refine-require-flags property
This allows `GsPluginJobFileToApp` to stop relying on that property from
`GsPluginJob`, which means we can soon remove it from the base class.
(See following commits.) This is a good idea because it’s only relevant
to a couple of `GsPluginJob` subclasses, and they can give it more
precise semantics.

Signed-off-by: Philip Withnall <pwithnall@gnome.org>

Helps: #1472