gs-plugin-job-url-to-app: Add a new allow-packages flag
This allows the job to stop querying the `GsPluginJob:refine-flags`
property for that flag, which means we can remove that property from the
base class (in the following commit). That’s tidier, because no other
subclasses use it.

It didn’t make complete sense for `GsPluginJobUrlToApp` to use
`GsPluginJob:refine-flags` either, as it was only querying the
`ALLOW_PACKAGES` flag; none of the other possible flags in the enum
affected the behaviour of `GsPluginJobUrlToApp`.

Signed-off-by: Philip Withnall <pwithnall@gnome.org>

Helps: #1472