gs-plugin-loader: Refactor data handling for processing jobs
Previously, the `GsPluginJob` was set as the `GTask`’s task data until
`job_process_cb()` was called, at which point it was replaced with the
start time of the job (for profiling).

The upcoming commits will need access to the `GsPluginJob` after
`job_process_cb()` is called, so this hack now needs to be replaced with
a more conventional task data struct.

Signed-off-by: Philip Withnall <pwithnall@gnome.org>