gs-plugin-job: Add GsPluginJob::event signal
This will eventually replace most (or all) uses of the
`GsPlugin::report-event` signal. It allows the `GsPluginJob` which
triggered an event to be associated with the event.

The following commits will hook it up to each of the `GsPluginJob`
subclasses, by modifying the `GsPlugin` vfuncs to accept a
`GsPluginEventCallback` argument, which the `GsPluginJob` subclass
provides. The `user_data` passed to that callback allows the
`GsPluginJob` to be associated with the event.

See the subsequent commits.

Signed-off-by: Philip Withnall <pwithnall@gnome.org>

Helps: #1472