packagekit: Explicitly use gs_plugin_report_event() for one event
It turns out we can’t eliminate `gs_plugin_report_event()` entirely, as
it’s needed for reporting events from a PackageKit download job which is
started in response to a change in prepared updates on the system,
rather than as part of any `GsPluginJob`.

That’s fine, and makes sense semantically.

Signed-off-by: Philip Withnall <pwithnall@gnome.org>

Helps: #1472