gs-plugin-job-file-to-app: Port to new event reporting API
See the previous commits for the explanation of the architecture.

By passing a `GsPluginEventCallback` into the job vfunc, we gain a
`user_data` argument which can be used to associate the `GsPluginJob`
with the `GsPluginEvent`.

In the long term, this can be used to access job data when reporting the
event to the user, and means we can eventually remove `GsPluginAction`,
as it duplicates the `GsPluginJob` subtypes.

Signed-off-by: Philip Withnall <pwithnall@gnome.org>

Helps: #1472