rpm-ostree: Ensure GsPluginEventCallbacks are called in the right thread
The rpm-ostree plugin does most of its work in a worker thread, so if we
call the event callback directly from that thread, the calling context
will need to do all sorts of locking to be safe.

Avoid that problem, and make the API more predictable and conformant to
GObject norms, by invoking the callback in the context which was the
thread-default when the original `_async` virtual method was called.

The event callback is invoked at `G_PRIORITY_DEFAULT` in the return
`GMainContext` of the `GTask`, and
`gs_rpm_ostree_task_return_error_with_gui()` is only ever called before
`g_task_return_*()` is called.

`g_task_return_*()` from a thread will invoke the task’s result function
at `G_PRIORITY_DEFAULT_IDLE` in the return `GMainContext` of the `GTask`.

Given the priority ordering of `GMainContext`, this guarantees that all
the event callbacks will be invoked by the task’s return `GMainContext`
before the task’s result function is. i.e. Before the
`GsPlugin.update_apps_async()` call (or whatever) asynchronously finishes.
The job is guaranteed to be alive across the duration of that async call.

Signed-off-by: Philip Withnall <pwithnall@gnome.org>

Helps: #1472