packagekit: Update some thread safety documentation
I was checking whether any of the `event_callback()`s were called in a
thread other than the main thread. One of them is, but that’s OK because
it only calls `gs_plugin_report_event()`, which is threadsafe already.

We could probably eventually simplify things by removing the locking
from `gs_plugin_report_event()` (and the signal handlers which are
connected to it), and instead emit this event in an idle callback on the
main thread — but this code will do for now. It’s safe and correct.

Signed-off-by: Philip Withnall <pwithnall@gnome.org>

Helps: #1472