gs-plugin-loader: Set GsPluginEvent:job on received events
Now that the infrastructure is in place to allow us to associate a
`GsPluginJob` with a `GsPluginEvent`, let’s do it.

This means that almost all `GsPluginEvent`s should now have their `job`
property set. The only events which don’t are those coming from the two
`gs_plugin_report_event()` calls, one of which happens in a background
PackageKit download operation, and the other happens on setup in the
flatpak plugin.

Signed-off-by: Philip Withnall <pwithnall@gnome.org>

Helps: #1472