gs-plugin-loader: Allow app to be specified explicitly when claiming an error
Rather than trying to extract the app from the `GsPluginJob`, allow the
caller to explicitly specify it to `gs_plugin_loader_claim_job_error()`.

This allows us to deal with several situations:
 - The job is created with multiple apps and none of them are directly
   relevant to an error reported at the end of the job.
 - The job is created with explicitly only one app, but the generic
   `GsPluginJob` code can’t know to return that one (compare with the
   previous case).
 - The job happens to be created with only one app, but it’s not
   directly relevant to the error reported at the end of the job.

If a job really needs to report an error about a specific app, then it
should be emitting a `GsPluginEvent` itself directly.

This also means we can now get rid of `gs_plugin_job_get_app()`, which
was a workaround for earlier API problems. See following commits.

Signed-off-by: Philip Withnall <pwithnall@gnome.org>

Helps: #1472