gs-repos-dialog: Use results of refine operation
The list of apps may change from the input to the result, as wildcard
apps are refined. So it’s safer to operate using the result list, rather
than the input list, after the job is finished.

This was spotted through code inspection, rather than in response to a
specific bug / set of symptoms.

Signed-off-by: Philip Withnall <pwithnall@gnome.org>