gs-plugin-job: Consistently set an error if no plugins support a job
Previously an error was only set for ‘core’ operations, and other jobs
would silently return as if they succeeded with an empty result set.

As discussed in
https://gitlab.gnome.org/GNOME/gnome-software/-/merge_requests/1266#note_1399606,
it would be better to always return an error, then the caller can decide
whether to show the error to the user or silently pretend the operation
succeeded.

I’ve not modified any of the call sites, as actually I think I *would*
want to see an error toast if any of these jobs failed due to no plugins
supporting them. We can always filter out some of the errors in future
on a case-by-case basis if they turn out to not be relevant. I wouldn’t
expect many people to see them, though, as all of these vfuncs are
provided by the core plugins (e.g. PackageKit, flatpak).

Signed-off-by: Philip Withnall <pwithnall@gnome.org>

Helps: #1472