gs-page: Stop magically choosing which cancellable to use
Let the caller decide, and otherwise don’t use a cancellable. This makes
the choice of cancellable a bit less opaque to callers.

I’ve modified all the callers of these functions to explicitly provide a
cancellable, so there are no functional changes in this commit.
Following commits will rework the choice of cancellable in the callers.

Signed-off-by: Philip Withnall <pwithnall@gnome.org>

Helps: #1472