gs-page: Remove magical cancellation of old jobs when uninstalling an app
This was needed when there was a limit to the number of operations in
the old threading model, but now it’s no longer helpful.

In particular, jobs can now operate on multiple apps, so cancelling any
job which references a particular app could cause (for example) the
ongoing update of multiple other apps to be cancelled. Instead, we
should rely on the underlying package manager to synchronise pending
operations correctly. That might result in an unnecessary update of an
app just before it’s uninstalled, but that’s not as bad as cancelling
the update of loads of other apps.

Signed-off-by: Philip Withnall <pwithnall@gnome.org>