plugins: Document the threading model of all other plugins
With this commit, every plugin now has its threading model documented,
which should make it a little bit easier to maintain them correctly in
future (e.g. paying attention to necessary locking of data).

As part of this, I’ve gone through and reviewed what each plugin does
for threading and locking, and removed some redundant code in the
process (see previous few commits).

I think we’re now in about the right place with plugin threading: only
the plugins which *need* to use threads (e.g. because they have to call
synchronous third party APIs, or because they do CPU intensive work) are
using threads, and all other plugins are running asynchronously in the
main thread.

So let’s take this opportunity to mark #1472 as fixed. There’s going to
be a long tail of cleanups which can be done on the back of this work,
but the main goal – reworking the threading model – is complete.
gnome-software has gone from running every vfunc invocation for every
plugin in a `GTask` thread pool thread (the availability of which is
limited); to running most things in the main thread, and explicitly
deferring some work to worker threads which are specific to the plugins
which need them.

In doing this, a lot of plugins have been able to drop their locking,
which should further improve performance.

The plugin API has also been modernised, batched (so operations can
happen on multiple apps at once, saving overhead), and moved towards
being introspectable (although that needs more work, e.g. see !2182).

Signed-off-by: Philip Withnall <pwithnall@gnome.org>
Fixes: #1472